+++ tests/single_card_tests/custom_ops/test_tilelang_dsv4_csa_indexer_kernel_fwd.py
@@ -17,7 +17,7 @@ import unittest
 import paddle
 import paddle.nn.functional as F
 
-from paddlefleet.ops.tilelang_dsv4.kernel.tilelang_csa_indexer_fwd import (
+from paddlefleet.tilelang_ops.kernel.tilelang_csa_indexer_fwd import (
     csa_indexer_topk_fwd_interface,
 )
